Define ThemePreference within MonoKnightApp
